+++ src/main/java/com/example/demo/config/SecurityConfig.java
@@ -1,6 +1,6 @@
 package com.example.demo.config;
 
-import com.example.demo.user.member.MemberSecurityService;
+import com.example.demo.user.member.service.MemberSecurityService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
