+++ src/main/java/com/example/demo/post/entity/Post.java
@@ -1,6 +1,6 @@
-package com.example.demo.post;
+package com.example.demo.post.entity;
 
-import com.example.demo.user.member.Member;
+import com.example.demo.user.member.entity.Member;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.data.annotation.CreatedDate;
