+++ src/main/java/com/example/demo/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
 package com.example.demo.post.repository;
 
-import com.example.demo.post.Post;
+import com.example.demo.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
