+++ src/main/java/com/example/demo/user/member/controller/MemberController.java
@@ -1,8 +1,7 @@
 package com.example.demo.user.member.controller;
 
-import com.example.demo.user.member.MemberCreateForm;
+import com.example.demo.user.member.form.MemberCreateForm;
 import com.example.demo.user.member.service.MemberService;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Controller;
