+++ src/main/java/com/example/demo/user/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.example.demo.user.member.repository;
 
-import com.example.demo.user.member.Member;
+import com.example.demo.user.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
