+++ src/main/java/com/example/demo/user/member/service/MemberSecurityService.java
@@ -1,5 +1,6 @@
-package com.example.demo.user.member;
+package com.example.demo.user.member.service;
 
+import com.example.demo.user.member.entity.Member;
 import com.example.demo.user.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
@@ -9,10 +10,7 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
